VACMS-20856: Moves FAQs unDrupaling to new file
