+++ src/data/queries/vamcSystemVaPolice.ts
@@ -16,6 +16,7 @@ import { Menu } from '@/types/drupal/menu'
 import { buildSideNavDataFromMenu } from '@/lib/drupal/facilitySideNav'
 import { getHtmlFromField } from '@/lib/utils/getHtmlFromField'
 import { QaSection as FormattedQaSection } from '@/types/formatted/qaSection'
+import { buildFaqs } from './ccFaqs'
 
 // Define the query params for fetching node--vamc_system_va_police.
 export const params: QueryParams<null> = () => {
@@ -70,32 +71,6 @@ export const formatter: QueryFormatter<
   VamcSystemVaPolice
 > = ({ entity, menu }) => {
   const formattedMenu = buildSideNavDataFromMenu(entity.path.alias, menu)
-  // Similarly, this formats centralized content FAQs to match what our QA components are expecting
-  const buildFaqs = (faqs) => {
-    const buildQuestionArray = (questions) => {
-      if (!questions) return []
-      return questions.map((question) => ({
-        id: question.target_id || null,
-        question: question.field_question[0]?.value || null,
-        answers: [
-          {
-            html: question.field_answer[0]?.field_wysiwyg[0]?.processed || null,
-          },
-        ],
-        header: question.label || null,
-      }))
-    }
-
-    return {
-      type: PARAGRAPH_RESOURCE_TYPES.QA_SECTION as FormattedQaSection['type'],
-      id: faqs.target_id,
-      header: faqs.fetched.field_section_header[0]?.value || null,
-      intro: faqs.fetched.field_section_intro[0]?.value || null,
-      displayAccordion:
-        Boolean(faqs.fetched.field_accordion_display[0]?.value) || false,
-      questions: buildQuestionArray(faqs.fetched.field_questions),
-    }
-  }
 
   return {
     ...entityBaseFields(entity),
